requirements.txt
# Python dependencies
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0

